Add task3 part3

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,5 @@
 from flask_restx import Namespace, Resource, fields
-
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.models import review
 from app.models.review import Review
 from app.services import facade
@@ -19,9 +19,21 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
+        current_user_id = get_jwt_identity()
         review_data = api.payload
+        place = facade.get_place(review_data['place_id'])
+        if not place:
+            return {'error': 'Place not found'}, 404
+        if str(place.owner.id) != str(current_user_id):
+            return {'message': 'You cannot review your own place'}, 400
+        all_reviews = facade.get_all_reviews()
+        for r in all_reviews:
+            if str(r.place.id) == str(review_data['place_id']) and str(r.user_id) == str(current_user_id):
+                return {'error': 'You have already reviewed this place'}, 400
+        review_data['user_id'] = current_user_id
         try:
             new_review = facade.create_review(review_data)
             if not new_review:
@@ -73,8 +85,15 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user_id = get_jwt_identity()
+        review = facade.get_review(review_id)
+        if not review:
+            return {'message': 'Review not found'}, 404
+        if str(review.user.id) != str(current_user_id):
+            return {'message': 'Unauthorized action'}, 400
         review_data = api.payload
         try:
             updated_review = facade.update_review(review_id, review_data)
@@ -86,9 +105,13 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user_id = get_jwt_identity()
         review = facade.get_review(review_id)
+        if str(review.user.id) != str(current_user_id):
+            return {'message': 'Unauthorized action'}, 400
         if not review:
             return {'message': 'Review not found'}, 404
         facade.delete_review(review_id)
